cleanup - rename library for ease, add more doc comments
