+++ app/examples/CMakeLists.txt
@@ -34,7 +34,7 @@ add_executable(poisson_test poisson_test.cpp)
 target_link_libraries(poisson_test deal.II::Interface)
 
 add_executable(run-factory run_factory.cpp)
-target_link_libraries(run-factory FunctionFactory)
+target_link_libraries(run-factory Factory)
 
 install(TARGETS dealii_function_test atan_test bit_test 
         # lagrange_test lagrange_time_test
