+++ src/Factory/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(Factory SHARED Function.cpp)
+target_include_directories(Factory PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+install(TARGETS Factory LIBRARY DESTINATION lib)
