+++ src/Factory/Function.cpp
@@ -2,7 +2,7 @@
 #include "Function.hpp"
 #include "Factory.hpp"
 
-namespace functionfactory {
+namespace factory {
 
 void Function::declare(const std::string& full_name, FunctionMaker* maker) {
   Factory::get().declare(full_name, maker);
